+++ src/bin/cli.js
@@ -12,6 +12,20 @@ const {
 } = require("../lib/dependency-tree-service");
 const { logger } = require("./logger");
 
+function printModuleList(dependencyTreeObject) {
+  if (dependencyTreeObject && dependencyTreeObject.length) {
+    logger.log(
+      `\n#################### Module list ####################\n\n${dependencyTreeObject
+        .map(e => `\t- ${e.moduleId}`)
+        .join("\n")}\n\n`
+    );
+  } else {
+    logger.warn(
+      "The module list is empty. Either the filters are too restrictive or there's no maven dependency:tree information in log file."
+    );
+  }
+}
+
 function main() {
   const args = getArgumentsObject();
   logger.level = args.debug ? "debug" : "info";
@@ -27,21 +41,39 @@ function main() {
   const dependencyTreeObject = getDependencyTreeObjectFromFile(
     args.logFilePath
   );
-  const dependencyTreeObjectFiltered = args.filter
-    ? filterDependencyTree(dependencyTreeObject, args.filter)
-    : dependencyTreeObject;
-  logger.debug(
-    `Filtered by ${args.filter}. Filtered ${dependencyTreeObjectFiltered.length} vs original ${dependencyTreeObject.length} `
+  const dependencyTreeObjectFiltered = filterDependencyTree(
+    dependencyTreeObject,
+    args.filter,
+    args.exclude
   );
-
-  const outputContent = dependencyTreeToString(dependencyTreeObjectFiltered);
-  fs.writeFileSync(path.join(".", pathToOutputFile), outputContent);
+  if (args.printModuleList) {
+    printModuleList(dependencyTreeObjectFiltered);
+  }
 
   logger.info(
-    `Execution output saved at ${pathToOutputFile}. Number of modules: ${
-      dependencyTreeObject ? dependencyTreeObject.length : 0
-    }`
+    `Filtered ${dependencyTreeObjectFiltered.length} module/s vs original ${
+      dependencyTreeObject.length
+    } module/s. ${
+      args.filter ? `Filtered by "${args.filter.join(", ")}"` : ""
+    }. ${args.exclude ? `Excluded by "${args.exclude.join(", ")}"` : ""}`
   );
+
+  if (!arguments.skipOutput) {
+    const outputContent = dependencyTreeToString(dependencyTreeObjectFiltered);
+    fs.writeFileSync(path.join(".", pathToOutputFile), outputContent);
+
+    logger.info(
+      `Execution output saved at ${pathToOutputFile}. Number of modules: ${
+        dependencyTreeObjectFiltered ? dependencyTreeObjectFiltered.length : 0
+      }`
+    );
+  } else {
+    logger.info(
+      `Execution output not saved due to arguments. Number of modules: ${
+        dependencyTreeObjectFiltered ? dependencyTreeObjectFiltered.length : 0
+      }`
+    );
+  }
 }
 
 if (require.main === module) {