+++ src/lib/dependency-tree-service.js
@@ -1,9 +1,36 @@
 const { outputSplitModule, separationBetweenModules } = require("./constants");
 const { createRegex } = require("../util/regex-util");
 
-function filterDependencyTree(dependencyTreeObject, filterRegExp) {
-  const regEx = createRegex(filterRegExp);
-  return dependencyTreeObject.filter(e => regEx.exec(e.info));
+/**
+ *
+ * @param {object} dependencyTreeObject
+ * @param {array} filterRegExpArray the array of regex expressions to filter by
+ * @param {array} excludeRegExpArray the array of regex expressions to exclude from
+ */
+function filterDependencyTree(
+  dependencyTreeObject,
+  filterRegExpArray,
+  excludeRegExpArray = []
+) {
+  const regExArray = filterRegExpArray
+    ? filterRegExpArray.map(filterRegExp => createRegex(filterRegExp))
+    : [];
+  const excludeRegExArray = excludeRegExpArray
+    ? excludeRegExpArray.map(filterRegExp => createRegex(filterRegExp))
+    : [];
+  return dependencyTreeObject
+    .filter(
+      dependentryTreeElement =>
+        regExArray
+          .map(regEx => regEx.exec(dependentryTreeElement.info))
+          .filter(e => e).length === regExArray.length
+    )
+    .filter(
+      dependentryTreeElement =>
+        excludeRegExArray
+          .map(regEx => regEx.exec(dependentryTreeElement.info))
+          .filter(e => e).length === 0
+    );
 }
 
 function dependencyTreeToString(dependencyTreeObject) {